requirements.txt
# environments
python-dotenv
torch

# models
pyannote.audio
whisper
faster-whisper

# eval
jiwer
matplotlib

# API
fastapi[standard]
